refactor(styles) - restructure components folder
