+++ src/components/AboutMe/index.ts
@@ -1,0 +1,1 @@
+export { default as AboutMeSection } from './AboutMeSection/AboutMeSection'
