+++ src/components/Projects/index.ts
@@ -1,0 +1,1 @@
+export { default as ProjectsSection } from './ProjectsSections/ProjectsSection'
