--- src/components/sections/index.ts
@@ -1,2 +1,0 @@
-export { default as AboutMe } from './AboutMe/AboutMe'
-export { default as ProjectsSection } from './Projects/ProjectsSection'
